# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warm House Platform API
  version: 1.0.0
  description: |
    REST API платформы Warm House для управления умным домом.

    Клиенты (веб/мобильное приложение) обращаются к **API Gateway** по HTTPS.
    Шлюз валидирует JWT-токен и проксирует запросы к внутренним микросервисам
    по gRPC. Данный документ описывает контракт **клиент ↔ шлюз**.

    **Асинхронные события** (Kafka + Protobuf) описаны в `schemas/asyncapi.yaml`.
servers:
  - url: https://api.warmhouse.io/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Локальная разработка
security:
  - BearerAuth: []
...
paths:
  /devices:
    post:
      tags:
        - Devices
      operationId: registerDevice
      summary: Зарегистрировать устройство
      description: |
        Регистрирует новое физическое устройство в платформе (онбординг).
        Тип устройства (`type_id`) должен существовать в каталоге `device_types`.

        После регистрации устройство получает статус `pending` — оно ещё
        не привязано к конкретному дому/комнате. Привязка выполняется отдельным
        запросом (`POST /devices/{deviceId}/bindings`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
            examples:
              thermostat:
                summary: Регистрация термостата
                value:
                  name: "Термостат кухня"
                  serial_number: "SN-2024-KT-001"
                  type_id: "a1b2c3d4-0000-0000-0000-000000000001"
                  protocol: MQTT
      responses:
        '201':
          description: Устройство успешно зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCreatedResponse'
              examples:
                thermostat:
                  summary: Термостат зарегистрирован
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Термостат кухня"
                    serial_number: "SN-2024-KT-001"
                    type_id: "a1b2c3d4-0000-0000-0000-000000000001"
                    protocol: MQTT
                    status: pending
                    registered_at: "2026-03-29T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Устройство с таким serial_number уже зарегистрировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: "Устройство с serial_number 'SN-2024-KT-001' уже существует"
                details: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT-токен, выданный Auth Service.
        Передаётся в заголовке: `Authorization: Bearer <token>`
...
  responses:
    BadRequest:
      description: Невалидные данные запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: "Поле 'serial_number' обязательно"
            details:
              field: serial_number
    Unauthorized:
      description: Отсутствует или невалиден JWT-токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: "Токен авторизации отсутствует или истёк"
            details: {}
    Forbidden:
      description: Нет прав на выполнение операции
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: "Недостаточно прав для управления этим устройством"
            details: {}
...
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: "Внутренняя ошибка сервера. Попробуйте позже."
            details: {}
  schemas:
    Error:
      type: object
      description: Стандартная структура ошибки
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: NOT_FOUND
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "Устройство не найдено"
        details:
          type: object
          additionalProperties: true
          description: Дополнительный контекст ошибки (зависит от типа ошибки)
          example: {}

    # --- Device schemas ---
    RegisterDeviceRequest:
      type: object
      required:
        - name
        - serial_number
        - type_id
        - protocol
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Читаемое название устройства
          example: "Термостат кухня"
        serial_number:
          type: string
          minLength: 1
          maxLength: 100
          description: Серийный номер с наклейки на устройстве (уникален в платформе)
          example: "SN-2024-KT-001"
        type_id:
          type: string
          format: uuid
          description: UUID типа устройства из каталога device_types
          example: "a1b2c3d4-0000-0000-0000-000000000001"
        protocol:
          type: string
          enum:
            - MQTT
            - MATTER
            - ZIGBEE
            - HTTP
          description: Протокол связи устройства с платформой
          example: MQTT
    DeviceCreatedResponse:
      type: object
      required:
        - id
        - name
        - serial_number
        - type_id
        - protocol
        - status
        - registered_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Термостат кухня"
        serial_number:
          type: string
          example: "SN-2024-KT-001"
        type_id:
          type: string
          format: uuid
          example: "a1b2c3d4-0000-0000-0000-000000000001"
        protocol:
          type: string
          enum:
            - MQTT
            - MATTER
            - ZIGBEE
            - HTTP
          example: MQTT
        status:
          type: string
          description: Статус онбординга устройства
          enum:
            - pending
            - active
            - inactive
          example: pending
        registered_at:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00Z"
...